requirements.txt
flask
flask-cors
sqlalchemy
pymysql
python-dotenv
flask-sqlalchemy
cryptography
flask-swagger-ui
requests
